+++ Database/RethinkDB/NoClash.hs
@@ -11,4 +11,5 @@ import Database.RethinkDB hiding (
   sum, map, mod, concatMap, (&&),
   not, (||), (/=), (<), (<=), (>), (>=), error, (==), filter,
   max, min,
-  zip, zipWith)
+  zip, zipWith,
+  floor, ceil, round)